pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-legal_auditor_env"
version = "0.1.0"
description = "Legal Auditor Env environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.8.4",
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    # install from github
    # "openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git",
    "fpdf2>=2.8.7",
    "matplotlib>=3.10.8",
    "openenv-core[core]>=0.2.1",
    # Environment-specific dependencies
    # Add all dependencies needed for your environment here
    # Examples:
    # "numpy>=1.19.0",
    # "torch>=2.0.0",
    # "gymnasium>=0.29.0",
    # "openspiel>=1.0.0",
    # "smolagents>=1.22.0,<2",
    "pandas>=2.3.3",
    "pymupdf>=1.27.2.2",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m legal_auditor_env.server.app
server = "legal_auditor_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["legal_auditor_env", "legal_auditor_env.server"]
package-dir = { "legal_auditor_env" = ".", "legal_auditor_env.server" = "server" }
